+++ src/main/java/team/yi/maven/plugin/config/GitlogPluginSettings.java
@@ -1,9 +1,0 @@
-package team.yi.maven.plugin.config;
-
-import team.yi.tools.semanticgitlog.config.GitlogSettings;
-
-public class GitlogPluginSettings extends GitlogSettings {
-    private static final long serialVersionUID = -857288928493905104L;
-
-    // TODO @Parameter
-}